Move endpoint validation to common part
