+++ central/externalbackups/plugins/s3/common/s3common_test.go
@@ -1,4 +1,4 @@
-package s3compatible
+package s3common
 
 import (
 	"testing"
