+++ central/externalbackups/plugins/s3/compatible/s3compatible.go
@@ -1,15 +1,12 @@
 package s3compatible
 
 import (
-	"net/url"
-
 	"github.com/pkg/errors"
 	"github.com/stackrox/rox/central/externalbackups/plugins"
 	s3common "github.com/stackrox/rox/central/externalbackups/plugins/s3/common"
 	"github.com/stackrox/rox/central/externalbackups/plugins/types"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/errorhelpers"
-	"github.com/stackrox/rox/pkg/urlfmt"
 )
 
 func init() {
@@ -39,19 +36,6 @@ func (c *s3compatibleConfigWrapper) GetUrlStyle() storage.S3URLStyle {
 
 func (c *s3compatibleConfigWrapper) GetEndpoint() string {
 	return c.integration.GetS3Compatible().GetEndpoint()
-}
-
-func (c *s3compatibleConfigWrapper) GetValidatedEndpoint() (string, error) {
-	return validateEndpoint(c.GetEndpoint())
-}
-
-func validateEndpoint(endpoint string) (string, error) {
-	// The aws-sdk-go-v2 package does not add a default scheme to the endpoint.
-	sanitizedEndpoint := urlfmt.FormatURL(endpoint, urlfmt.HTTPS, urlfmt.NoTrailingSlash)
-	if _, err := url.Parse(sanitizedEndpoint); err != nil {
-		return "", errors.Wrapf(err, "invalid URL %q", endpoint)
-	}
-	return sanitizedEndpoint, nil
 }
 
 func (c *s3compatibleConfigWrapper) GetRegion() string {
